expressions are working for floating numbers

+++ src/library/Evaluator.java
@@ -5,64 +5,66 @@ import java.util.List;
 
 public class Evaluator {
 
-    private int evaluate(int num1, String operator, int num2) {
-        int result = 0;
+    private double evaluate(double num1, String operator, double num2) {
+        double result = 0.0;
         if (operator.equals("+")) result = num1 + num2;
         if (operator.equals("-")) result = num1 - num2;
         if (operator.equals("*")) result = num1 * num2;
         if (operator.equals("/")) result = num1 / num2;
-        if (operator.equals("^")) result = (int) Math.pow(num1, num2);
+        if (operator.equals("^")) result = (float) Math.pow(num1, num2);
         return result;
     }
-    public int evaluateExpression(String expression) throws Exception {
-        if (!expression.contains(" ")) return Integer.parseInt(expression);
-        if (expression.contains("(")){
+
+    public double evaluateExpression(String expression) throws Exception {
+        if (!expression.contains(" ")) return Double.parseDouble(expression);
+        if (expression.contains("(")) {
             String res = evaluateWithBrackets(expression);
             return evaluateExpression(res);
         }
         String[] data = expression.split(" ");
-        List<Integer> operands = new ArrayList<Integer>();
+        List<Double> operands = new ArrayList<Double>();
         List<String> operators = getOperators(data, operands);
-        int result = evaluateMultipleOperations(operands, operators);
+        double result = evaluateMultipleOperations(operands, operators);
         return result;
     }
 
     private String evaluateWithBrackets(String expression) throws Exception {
         StringBuffer sb = new StringBuffer(expression);
-        int startIndex = 0,endIndex = 0;
-        for (int i = 0; i < expression.length() ; i++) {
-            if(expression.charAt(i) == '(')
-                startIndex = i;
-            else if(expression.charAt(i) == ')'){
+        int startIndex = 0, endIndex = 0;
+        for (int i = 0; i < expression.length(); i++) {
+            if (expression.charAt(i) == '(') startIndex = i;
+            else if (expression.charAt(i) == ')') {
                 endIndex = i;
-                 break;
+                break;
             }
         }
         String expressionInBrackets = expression.substring(startIndex + 1, endIndex);
-        int res = evaluateExpression(expressionInBrackets);
-        sb.replace(startIndex, endIndex + 1, Integer.toString(res));
+        double res = evaluateExpression(expressionInBrackets);
+        sb.replace(startIndex, endIndex + 1, Double.toString(res));
         return sb.toString();
     }
-    private int evaluateMultipleOperations(List<Integer> operands, List<String> operators) throws Exception {
+
+    private double evaluateMultipleOperations(List<Double> operands, List<String> operators) throws Exception {
         Evaluator e = new Evaluator();
-        int result = e.evaluate(operands.get(0), operators.get(0), operands.get(1));
+        double result = e.evaluate(operands.get(0), operators.get(0), operands.get(1));
         for (int i = 1; i < operators.size(); i++) {
             result = e.evaluate(result, operators.get(i), operands.get(i + 1));
         }
         return result;
     }
-    private List<String> getOperators(String[] data, List<Integer> operands) {
+
+    private List<String> getOperators(String[] data, List<Double> operands) {
         List<String> operators = new ArrayList<String>();
         List list = new ArrayList();
         for (String s : data) {
             try {
-                int number = Integer.parseInt(s);
+                Double number = Double.parseDouble(s);
                 operands.add(number);
             } catch (Exception ex) {
                 operators.add(s);
             }
         }
-        if(operands.size()-1 == operators.size())
+        if (operands.size() - 1 == operators.size())
             return operators;
         list.add("wrong input");
         return list;
